[CL-1941] Move all shared components to app/components/ContentBuilder
