+++ front/app/components/ContentBuilder/TopBar/GoBackButton.tsx
