+++ front/app/components/ContentBuilder/TopBar/messages.ts
@@ -1,0 +1,12 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  preview: {
+    id: 'app.containers.AdminPage.ProjectDescription.preview',
+    defaultMessage: 'Preview',
+  },
+  contentBuilderSave: {
+    id: 'app.containers.ContentBuilder.Save',
+    defaultMessage: 'Save',
+  },
+});
