+++ front/app/modules/commercial/content_builder/admin/components/ContentBuilderTopBar/messages.ts
@@ -5,17 +5,9 @@ export default defineMessages({
     id: 'app.containers.ContentBuilder.PageTitle',
     defaultMessage: 'Project description',
   },
-  contentBuilderSave: {
-    id: 'app.containers.ContentBuilder.Save',
-    defaultMessage: 'Save',
-  },
   viewPublicProject: {
     id: 'app.containers.AdminPage.ProjectDescription.viewPublicProject',
     defaultMessage: 'View project',
-  },
-  preview: {
-    id: 'app.containers.AdminPage.ProjectDescription.preview',
-    defaultMessage: 'Preview',
   },
   viewProject: {
     id: 'app.containers.AdminPage.ProjectDescription.viewProject',
